+++ alsongDalsong/alsongDalsong/alsongDalsong/Sources/Extensions/UIViewController+present.swift
@@ -4,8 +4,4 @@ extension UIViewController {
     func presentAlert(_ viewcontrollerToPresent: ASAlertController) {
         present(viewcontrollerToPresent, animated: true, completion: nil)
     }
-
-    func presentLoadingView(_ viewcontrollerToPresent: ASAlertController) {
-        present(viewcontrollerToPresent, animated: true, completion: nil)
-    }
 }
