+++ alsongDalsong/alsongDalsong/alsongDalsong/Sources/Views/Onboarding/OnboardingViewController.swift
@@ -225,7 +225,7 @@ extension OnboardingViewController {
 
 extension OnboardingViewController {
     private func showRoomNumerInputAlert() {
-        let alert = ASAlertController(
+        let alert = InputAlertController(
             titleText: .joinRoom,
             textFieldPlaceholder: .roomNumber,
             isUppercased: true
@@ -236,12 +236,12 @@ extension OnboardingViewController {
     }
 
     private func showRoomFailedAlert(_ error: Error) {
-        let alert = ASAlertController(titleText: .error(error))
+        let alert = SingleButtonAlertController(titleText: .error(error))
         presentAlert(alert)
     }
 
     private func showCreateRoomLoading() {
-        let alert = ASAlertController(
+        let alert = LoadingAlertController(
             progressText: .joinRoom,
             load: { [weak self] in
                 try await self?.setNicknameAndCreateRoom()
@@ -250,7 +250,7 @@ extension OnboardingViewController {
                 self?.showRoomFailedAlert(error)
             }
         )
-        presentLoadingView(alert)
+        presentAlert(alert)
     }
 }
